+++ remote/src/androidTest/java/com/universodoandroid/starwarsjetpack/remote/ListPeopleUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.universodoandroid.starwarsjetpack.remote
 
 import com.nhaarman.mockitokotlin2.mock
-import com.universodoandroid.starwarsjetpack.remote.api.PeopleApiDataSource
+import com.universodoandroid.starwarsjetpack.remote.people.api.PeopleApiDataSource
 import org.junit.Before
 import org.junit.Test
