+++ remote/src/androidTest/java/com/universodoandroid/starwarsjetpack/remote/observers/FakePeopleRemoteDataSource.kt
@@ -1,13 +1,14 @@
 package com.universodoandroid.starwarsjetpack.remote.observers
 
-import com.universodoandroid.starwarsjetpack.remote.api.PeopleApiDataSource
+import com.universodoandroid.starwarsjetpack.remote.people.api.PeopleApiDataSource
 import com.universodoandroid.starwarsjetpack.remote.data.PeopleReponseDataFactory
-import com.universodoandroid.starwarsjetpack.remote.remote.people.response.PeopleResponse
+import com.universodoandroid.starwarsjetpack.remote.people.data.response.PeopleResponse
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
 import io.reactivex.Observable
 
-class FakePeopleRemoteDataSource : PeopleApiDataSource {
+class FakePeopleRemoteDataSource :
+    PeopleApiDataSource {
 
     override fun loadPeople(page: Int): Flowable<PeopleResponse> {
         return Observable.create<PeopleResponse> {
