+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/di/RemoteModules.kt
@@ -1,8 +1,8 @@
 package com.universodoandroid.starwarsjetpack.remote.di
 
 import com.universodoandroid.starwarsjetpack.remote.BuildConfig
-import com.universodoandroid.starwarsjetpack.remote.api.ApiDataSource
-import com.universodoandroid.starwarsjetpack.remote.remote.people.di.getPeopleRemoteModules
+import com.universodoandroid.starwarsjetpack.remote.ApiDataSource
+import com.universodoandroid.starwarsjetpack.remote.people.data.di.getPeopleRemoteModules
 import org.koin.dsl.module
 
 internal val apiDataSource = module {
