+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/people/data/PeopleRemoteDataSource.kt
@@ -1,0 +1,9 @@
+package com.universodoandroid.starwarsjetpack.remote.people.data
+
+import com.universodoandroid.starwarsjetpack.remote.people.data.response.PeopleResponse
+import io.reactivex.Flowable
+
+
+internal interface PeopleRemoteDataSource {
+    fun loadPeoplePerPage(page: Int): Flowable<PeopleResponse>
+}
