+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/planet/api/PlanetApiDataSource.kt
@@ -1,6 +1,6 @@
-package com.universodoandroid.starwarsjetpack.remote.api
+package com.universodoandroid.starwarsjetpack.remote.planet.api
 
-import com.universodoandroid.starwarsjetpack.remote.remote.planet.response.PlanetResponse
+import com.universodoandroid.starwarsjetpack.remote.planet.data.response.PlanetResponse
 import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Query
