+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/planet/data/PlanetRemoteDataSource.kt
@@ -1,6 +1,6 @@
-package com.universodoandroid.starwarsjetpack.remote.remote.planet
+package com.universodoandroid.starwarsjetpack.remote.planet.data
 
-import com.universodoandroid.starwarsjetpack.remote.remote.planet.response.PlanetResponse
+import com.universodoandroid.starwarsjetpack.remote.planet.data.response.PlanetResponse
 
 interface PlanetRemoteDataSource {
     fun loadPlanets(page: Int, onSuccess: (result: PlanetResponse) -> Unit, onError: (Throwable) -> Unit)
