+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/planet/data/response/PlanetResponse.kt
@@ -1,0 +1,8 @@
+package com.universodoandroid.starwarsjetpack.remote.planet.data.response
+
+data class PlanetResponse(
+    val count: Int,
+    val next: String,
+    val previous: Any,
+    val planets: List<Planet>
+)
